Set up an initializer to read AWS creds from Vault
Vault configuration can be set via `VAULT_ADDR` and `VAULT_TOKEN` - falls back to local env vars if Vault isn't present.

As an aside this switches to STS credentials using a session token,
which _ought_ to work
(https://aws.amazon.com/blogs/security/a-new-and-standardized-way-to-manage-credentials-in-the-aws-sdks/)
but could probably use testing before going straight to prod.

Should be able to test once https://github.com/ello/infrastructure/pull/48 is built/up.